+++ include/rom.h
@@ -19,7 +19,9 @@ rom_format_t rom_detect_format(const uint8_t *data, size_t size, const char *pat
 /* Parse SNES cartridge header and populate rom_t struct */
 int rom_parse_header(zsnes_emu_t *emu);
 
-/* De-interleave SMC format ROM to linear layout */
-int rom_deinterleave(uint8_t *data, size_t size);
+/* De-interleave SMC format ROM to linear layout.
+ * Returns the new payload size after any copier header is removed.
+ */
+size_t rom_deinterleave(uint8_t *data, size_t size);
 
 #endif /* ROM_H */
